+++ server/lib/configuration/channel/init.ts
@@ -106,7 +106,7 @@ async function initChannelConfiguration (options: RegisterServerOptions): Promis
           // FIXME: this piece of code should not be in this file (nothing to do with initChannelConfiguration,
           //   but will be more efficient to add here, as we already tested hasChat).
           // Note: no need to await here, would only degrade performances.
-          updateProsodyRoom(options, video.channelId, video.uuid, {
+          updateProsodyRoom(options, video.uuid, {
             name: video.name
           }).then(
             () => {},
@@ -133,7 +133,7 @@ async function initChannelConfiguration (options: RegisterServerOptions): Promis
       if (settings['prosody-room-type'] === 'channel') {
         const jid = 'channel.' + channel.id.toString()
         // Note: no need to await here, would only degrade performances.
-        updateProsodyRoom(options, channel.id, jid, {
+        updateProsodyRoom(options, jid, {
           name: channel.displayName
         }).then(
           () => {},